Allow passing of MarketPlaceClient instance during app instantiation. (#35)
Instead of passing through client constructor arguments, we enable users
to instantiate the client and then pass that directly. This is an
improvement to the composition model as it does not require the
replication of client constructor arguments within the app constructor
function and also enables users to re-use an existing client instance.